Исправление расположения файлов. Исритических ошибок. Представление

+++ src/main.ts
@@ -1,25 +1,25 @@
 import "./scss/styles.scss";
-import { Basket } from "./components/base/models/basket";
-import { Products } from "./components/base/models/products";
-import { Customer } from "./components/base/models/customer";
+import { Basket } from "./components/models/basket";
+import { Products } from "./components/models/products";
+import { Customer } from "./components/models/customer";
 import { API_URL } from "./utils/constants";
 import { Api } from "./components/base/Api";
 import { ApiClient } from "./components/base/ApiClient";
-import { Gallery } from "./components/base/view/Gallery";
+import { Gallery } from "./components/view/Gallery";
 import { cloneTemplate, ensureElement } from "./utils/utils";
 import { EventEmitter } from "./components/base/Events";
 
-import { CardCatalog } from "./components/base/view/card/CardCatalog";
+import { CardCatalog } from "./components/view/card/CardCatalog";
 import { IProduct } from "./types/IProduct";
-import { CardPreview } from "./components/base/view/card/CardPreview";
-import { Modal } from "./components/base/view/Modal";
-import { Header } from "./components/base/view/Header";
-import { basketList } from "./components/base/view/BasketList";
-import { CardBasket } from "./components/base/view/card/CardBasket";
-import { Order } from "./components/base/view/Forms/Order";
-import { Contacts } from "./components/base/view/Forms/Contacts";
+import { CardPreview } from "./components/view/card/CardPreview";
+import { Modal } from "./components/view/Modal";
+import { Header } from "./components/view/Header";
+import { basketList } from "./components/view/BasketList";
+import { CardBasket } from "./components/view/card/CardBasket";
+import { Order } from "./components/view/Forms/Order";
+import { Contacts } from "./components/view/Forms/Contacts";
 import { IOrderRequest } from "./types/IOrderRequest";
-import { Success } from "./components/base/view/Success";
+import { Success } from "./components/view/Success";
 
 const events = new EventEmitter();
 const productsModel = new Products(events);
@@ -33,6 +33,7 @@ const basket = new basketList(events, cloneTemplate("#basket"));
 const order = new Order(events, cloneTemplate("#order"));
 const contacts = new Contacts(events, cloneTemplate("#contacts"));
 const success = new Success(events, cloneTemplate("#success"));
+const preview = new CardPreview(cloneTemplate('#card-preview'));
 
 const api = new Api(API_URL);
 const apiClient = new ApiClient(api);
@@ -49,14 +50,19 @@ apiClient
 events.on("products:change", () => {
   const cards = productsModel.getItems().map((item) => {
     const card = new CardCatalog(cloneTemplate("#card-catalog"), {
-      onClick: () => events.emit("product:select", item),
+      onClick: () => productsModel.saveCard(item),
     });
     return card.render(item);
   });
   gallery.render({ catalog: cards });
 });
 
-events.on("product:select", (element: IProduct) => {
+
+events.on("product:select", () => {
+  const element = productsModel.getCard();
+  if (!element) {
+    return;
+  }
   const isSelected = basketModel.hasItem(element.id);
   const buttonText =
     element.price == null
@@ -64,52 +70,52 @@ events.on("product:select", (element: IProduct) => {
       : isSelected
       ? "Удалить из корзины"
       : "Купить";
-  const preview = new CardPreview(cloneTemplate("#card-preview"), {
-    onClick: () => events.emit("product:chosen", element),
-  });
+  
+  preview.action = { onClick: () => events.emit("product:chosen", element) };
   modal.content = preview.render({
     ...element,
     buttonText: buttonText,
-    buttonDisabled: element.price == null,
+    disabled: element.price == null,
   });
   modal.open();
 });
 
-events.on("modal:close", () => {
-  modal.close();
-});
-
 events.on("basket:change", () => {
   header.counter = basketModel.getAmount();
+  const items = basketModel.getItems().map((item, index) => {
+    const card = new CardBasket(cloneTemplate("#card-basket"), {
+      onClick: () => {
+       { events.emit("basket:deleteItem", item), events.emit('basket:open') };
+      },
+    });
+    return card.render({ ...item, index: index + 1 });
+  });
+  basket.render({
+    price: basketModel.getTotalPrice(),
+    list: items,
+    buttonDisabled: basketModel.getAmount() == 0,
+  });
 });
 
-events.on("product:chosen", (element: IProduct) => {
+events.on("product:chosen", () => {
+  const element = productsModel.getCard();
+  if (!element) {
+    return;
+  }
   basketModel.hasItem(element.id)
     ? basketModel.deleteItem(element)
     : basketModel.addItem(element);
   modal.close();
 });
 
-events.on("basket:render", () => {
-  const items = basketModel.getItems().map((item, index) => {
-    const card = new CardBasket(cloneTemplate("#card-basket"), {
-      onClick: () => {
-        events.emit("basket:deleteItem", item);
-      },
-    });
-    return card.render({ ...item, index: index + 1 });
-  });
-  modal.content = basket.render({
-    price: basketModel.getTotalPrice(),
-    list: items,
-    buttonDisabled: basketModel.getAmount() == 0,
-  });
+events.on("basket:open", () => {
+  modal.content = basket.render();
   modal.open();
 });
 
 events.on("basket:deleteItem", (element: IProduct) => {
   basketModel.deleteItem(element);
-  events.emit("basket:render");
+  modal.close();
 });
 
 events.on("basket:order", () => {
@@ -128,22 +134,29 @@ events.on("order:address", (value: { address: string }) => {
 });
 
 events.on("customer:change", () => {
-  order.payment = customerModel.getCustomerData().payment;
-  order.addres = customerModel.getCustomerData().address;
-  contacts.email = customerModel.getCustomerData().email;
-  contacts.phone = customerModel.getCustomerData().phone;
-  order.error =
-    customerModel.validate().address || customerModel.validate().payment || "";
-  contacts.error =
-    customerModel.validate().email || customerModel.validate().phone || "";
+  const data = customerModel.getCustomerData();
+  order.payment = data.payment;
+  order.addres = data.address;
+  contacts.email = data.email;
+  contacts.phone = data.phone;
+  const errors = customerModel.validate()
+  const orderError = [errors.payment, errors.address]
+  .filter(value => value)
+  .join('; ');
+  const orderDisable =
+  customerModel.validate().address != undefined ||
+  customerModel.validate().payment != undefined;
+  
+  order.render({error: orderError, disable: orderDisable});
 
-  order.disable =
-    customerModel.validate().address != undefined ||
-    customerModel.validate().payment != undefined;
-
-  contacts.disable =
+  const contactsError = [errors.email, errors.phone]
+  .filter(value => value)
+  .join('; ');
+  const contactsDisable =
     customerModel.validate().email != undefined ||
     customerModel.validate().phone != undefined;
+
+  contacts.render({error: contactsError, disable: contactsDisable});
 });
 
 events.on("order:submit", () => {
@@ -161,11 +174,9 @@ events.on("contacts:phone", (value: { phone: string }) => {
 });
 
 events.on("contacts:submit", async () => {
+
   const order: IOrderRequest = {
-    payment: customerModel.getCustomerData().payment,
-    email: customerModel.getCustomerData().email,
-    phone: customerModel.getCustomerData().phone,
-    address: customerModel.getCustomerData().address,
+    ...customerModel.getCustomerData(), 
     items: basketModel.getItems().map((item) => item.id),
     total: basketModel.getTotalPrice(),
   };
